bunch of stuff
I added audio player objects, as well as bunch of audio. It can either play one sound when interacting with sprite, or loop a song (or ambience) when loading a scene.
I also added a preload method that will load all the images so that it doesnt load when playing the game and is seemless. It requiers php tho